refactor: enhance styling of StepInput
