+++ src/Component/FunctionUI/StepInput/StepInput.less
@@ -1,0 +1,7 @@
+.gs-step-input {
+  display: flex;
+
+  .number-expression-input {
+    flex: 1;
+  }
+}
